+++ internal/runner/runner.go
@@ -3,6 +3,7 @@ package runner
 import (
 	"encoding/json"
 	"io"
+	"os"
 	"strconv"
 
 	"github.com/psampaz/go-mod-outdated/internal/mod"
@@ -10,8 +11,10 @@ import (
 	"github.com/olekukonko/tablewriter"
 )
 
+var OsExit = os.Exit
+
 // Run converts the the json output of go list -u -m -json all to table format
-func Run(in io.Reader, out io.Writer, update, direct bool) error {
+func Run(in io.Reader, out io.Writer, update, direct, exitWithNonZero bool) error {
 	var modules []mod.Module
 	dec := json.NewDecoder(in)
 
@@ -21,13 +24,26 @@ func Run(in io.Reader, out io.Writer, update, direct bool) error {
 
 		if err != nil {
 			if err == io.EOF {
-				renderTable(out, mod.FilterModules(modules, update, direct))
+				filteredModules := mod.FilterModules(modules, update, direct)
+				renderTable(out, filteredModules)
+				if hasOutdated(filteredModules) && exitWithNonZero {
+					OsExit(1)
+				}
 				return nil
 			}
 			return err
 		}
 		modules = append(modules, v)
 	}
+}
+
+func hasOutdated(filteredModules []mod.Module) bool {
+	for m := range filteredModules {
+		if filteredModules[m].HasUpdate() {
+			return true
+		}
+	}
+	return false
 }
 
 func renderTable(writer io.Writer, modules []mod.Module) {